Fixed some more inconsistencies

+++ projects/miopen/src/solver/conv/conv_ck_igemm_fwd_v6r1_dlops_nchw.cpp
@@ -133,7 +133,7 @@ PerformanceConvCkIgemmFwdV6r1DlopsNchw
 ConvCkIgemmFwdV6r1DlopsNchw::GetDefaultPerformanceConfig(const ExecutionContext& ctx,
                                                          const ProblemDescription& problem) const
 {
-    for(int i = 0; static_cast<size_t>(i) <
+    for(size_t i = 0U; i <
                    ck::driver::ConvIgemmFwdV6r1DlopsNchwKcyxNkhw::GetTunableList().size();
         ++i)
     {
